% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bigram_prob}
\alias{bigram_prob}
\title{A matrix containing the bigram probabilities for each letter pair.}
\format{
An object of class \code{matrix} with 26 rows and 26 columns.
}
\usage{
bigram_prob
}
\description{
A matrix containing the bigram probabilities for each letter pair.
}
\keyword{datasets}
