% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_gibber.R
\name{is_gibberish}
\alias{is_gibberish}
\title{Title}
\usage{
is_gibberish(text, threshold = threshold)
}
\arguments{
\item{text}{character vector}

\item{threshold}{numeric}
}
\value{

}
\description{
Title
}
